+++ command/clistate/state.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/hashicorp/errwrap"
 	multierror "github.com/hashicorp/go-multierror"
-	"github.com/hashicorp/terraform/helper/slowmessage"
+	"github.com/hashicorp/terraform/internal/helper/slowmessage"
 	"github.com/hashicorp/terraform/states/statemgr"
 	"github.com/mitchellh/cli"
 	"github.com/mitchellh/colorstring"
