--- helper/README.md
@@ -1,9 +1,0 @@
-# Legacy Helper Libraries
-
-The packages in this directory are all legacy code. Some of them are legacy
-because they are now maintained in
-[the Terraform SDK](https://github.com/hashicorp/terraform-plugin-sdk),
-while others are just obsolete codepaths that we intend to migrate away
-from over time.
-
-Avoid using functions from packages under `helper/` in new projects.
