+++ internal/helper/wrappedreadline/wrappedreadline_unix.go
@@ -6,7 +6,7 @@ import (
 	"syscall"
 	"unsafe"
 
-	"github.com/hashicorp/terraform/helper/wrappedstreams"
+	"github.com/hashicorp/terraform/internal/helper/wrappedstreams"
 )
 
 // getWidth impl for Unix
