+++ internal/legacy/helper/hashcode/hashcode.go
