+++ internal/legacy/helper/hashcode/hashcode_test.go
