+++ internal/legacy/helper/schema/field_reader_config_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/hashicorp/terraform/configs/hcl2shim"
-	"github.com/hashicorp/terraform/helper/hashcode"
+	"github.com/hashicorp/terraform/internal/legacy/helper/hashcode"
 	"github.com/hashicorp/terraform/internal/legacy/terraform"
 )
