+++ internal/legacy/helper/schema/set.go
@@ -8,7 +8,7 @@ import (
 	"strconv"
 	"sync"
 
-	"github.com/hashicorp/terraform/helper/hashcode"
+	"github.com/hashicorp/terraform/internal/legacy/helper/hashcode"
 )
 
 // HashString hashes strings. If you want a Set of strings, this is the
